On sign in, redirect back to protected route
When a user who hadn't signed in tried to access a protected route,
they were always redirected back to the home page. There could be
property owners who have bookmarked a protected (private) route and
would prefer to be redirected back to that route when they successfully
sign in.

Adding the feature mentioned above by storing the originally requested
route's path and query parameters in redirect state. When the user is
redirected to sign in page and signs in, the original route is extracted
and the user is redirected back to it. If no state exists, such as when
sign in was accessed from a public page, it will default to redirect the
user to home page.

Additionally updated going back to the previous page to be done through
React Router's history.