Sending notifications from leaving page
Initially I created a separate notification component that allows them
to be shown by being extracted from query parameters. To show them, a
component had to redirect to another and pass specific message query
parameters and the receiving component would have to call the
notification object to display the message.

I found out that it is possible to directly display the message from a
component and immediately redirect to a new component, but the message
will still be displayed.

Using the updated way of displaying notification messages and removed
no longer used methods.