Created message details page
Created a page that shows a selected messages details. First it changes
the message's status to "read", then finds the associated property. It
displays the following information:
    - The associated properties tile;
    - The message sender's email and phone (if specified);
    - The message text.

This page supports the following actions:
    - If the message is not archived, allows to archive it;
    - If the message is archived, allows to unarchive it;
    - Deleting the message, with a confirmation dialog for acciedental
    clicks.

While the component is getting the data from the server, it displays a
loading skeleton and if the message isn't found, displays a warning
message.