Display the number of unread messages
Display the number of unread messages in a badge in the header. When a
new message is received, display notification. Querying for new messages
is done on page changes (when header is updated), so that socket
connection doesn't need to be implemented. This can cause slight delays,
but as long as the users are active in the website, they will quickly
be notified of all new messages.