Added documentation using JSDoc and better-docs
The code had been documented throughout the development process, but
until now there was no way to view this documentation.

Added JSDoc package to generate the documentation. Added better-docs
plugin which allows the documentation to be oranised by components and
provides a more readable theme for React applications.

Updated all the existing components to include the @component annotation
and added descriptions to property types. Fixed some copy-paste issues
which resulted in inaccurate documentation for some components.