Positioning hamburger menu for mobile devices
When the screen's width becomes small enough that the navigation bar
elements are no longer visible, an expandable hamburger menu appears.
By default it was positioned right after the logo, but this meant that
on mobile devices it appeared around the middle in the header. This was
counter-intuitive as users are expecting the menu to be at either of the
corners.

Updating LESS (CSS) style to position this element at the right side and
reduced the header's right padding, so that there isn't a huge gap after
the menu.