Added sign in tests and test setup
Created a test setup file to manually mock a method which has not been
implemented in JSDOM. Configured jest to call this file before any
testing.

Created tests for the Sign In component, testing successful and failed
sign in actions.

Updated Sign In component to show a message on successful sign in, which
can be used to wait for the asynchronous API call to occur by waiting
for the message to appear.