BugFix: Price parser didn't work when built
Price parser for Ant Design's InputNumber didn't allow the value to be
updated when the app was built for production, while it was working when
the app was served in development mode.

Updated the parser to handle more edge cases, including unexpected input
types.

Also fixed a memory leak in Navbar, by putting the code that contacts
the API inside a useEffect hook.